pyproject.toml
[build-system]
requires = ["hatchling", "pybind11", "setuptools~=65.5.0"]
build-backend = "hatchling.build"

[project]
name = "biobridge"
version = "0.2.0"
authors = [
    { name = "Okerew", email = "okerewgroup@proton.me" }
]
description = "A library for simulating all biology related things, analyzing them and visualizing them."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
dependencies = [
    "numpy~=1.26.4",
    "matplotlib~=3.9.2",
    "py3Dmol~=2.3.0",
    "opencv-python~=4.10.0.84",
    "mediapipe~=0.10.15",
    "scikit-learn~=1.5.1",
    "scipy~=1.14.1",
    "imageio~=2.35.1",
    "scikit-image~=0.24.0",
    "networkx~=3.3",
    "biopython~=1.84",
    "requests~=2.32.3",
    "pyusb~=1.2.1",
    "opcua~=0.98.13",
    "pyserial~=3.5",
    "Flask~=3.0.3",
    "pandas~=2.2.2",
    "okrolearn~=0.2.7",
    "pytesseract~=0.3.13",
    "pillow~=10.4.0",
    "pybind11~=2.13.5"
]

[project.urls]
Documentation = "https://github.com/Okerew/biobridge#readme"
Issues = "https://github.com/Okerew/biobridge/issues"
Source = "https://github.com/Okerew/biobridge"

[tool.hatch.version]
path = "biobridge/__about__.py"

[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/biobridge tests}"

[tool.coverage.run]
source_pkgs = ["biobridge", "tests"]
branch = true
parallel = true
omit = [
  "biobridge/__about__.py",
]

[tool.coverage.paths]
biobridge = ["biobridge", "*biobridge"]
tests = ["tests", "*/biobridge/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.wheel]
packages = ["biobridge"]

